requirements.txt
langchain 
langchain-community 
langchain-openai 
langchain-milvus 
pymilvus 
pypdf 
python-dotenv